Corrected StreamThreadTest, and moved tests for ProcessorTopologyTestDriver into ProcessorTopologyTest.
